refs #46606 add some helper args and debug output

+++ 22dakika-shortcodes/generate_oyuncu_list_from_json.php
@@ -29,7 +29,15 @@ $now = $now->format('Y-m-d H:i:s');
 $debug = in_array('-d', $argv);
 $dry_run = in_array('-n', $argv);
 $force = in_array('-f', $argv);
+$published_only = in_array('-p', $argv);
 
+function print_debug($str) {
+  global $debug;
+
+  if ($debug) {
+    echo $str . "\n";
+  }
+}
 
 if (!$force) {
   $date = get_option("oyuncu_listesi_son_tarih");
@@ -43,6 +51,9 @@ if (!$force) {
   }
 }
 
+$working_on_str = $force ? 'never forever' : $date;
+print_debug("Working on DB data after " . $working_on_str);
+
 $posts_table = $wpdb->prefix . "posts";
 // $sql = "SELECT post_content FROM $posts_table WHERE post_modified_gmt >= '$date'";
 $sql = "SELECT post_content FROM $posts_table WHERE post_content LIKE '%[oyuncu]%'";
@@ -51,12 +62,14 @@ if (!$force) {
   $sql .= " AND post_modified_gmt >= '$date'";
 }
 
+if ($published_only) {
+  $sql .= " AND post_status = 'publish'";
+}
+
 $post_list = $wpdb->get_results($sql);
 $oyuncu_listesi = array();
 
-if ($debug) {
-  echo "\n DB fetch finished, parsing content\n\n";
-}
+print_debug("DB fetch finished, parsing content");
 
 foreach ($post_list as $post) {
   $content = $post->post_content;
@@ -75,43 +88,44 @@ foreach ($post_list as $post) {
 $file_uri = plugin_dir_path(__FILE__) . '/oyuncu_listesi.json';
 $json_liste = array();
 
-if (file_exists($file_uri)) {
-  $fp = fopen($file_uri, 'r') or die('could not open file for reading: ' . $file_uri);
-  $json_liste = json_decode(fread($fp, filesize($file_uri)), true);
-  fclose($fp);
-}
+print_debug("reading file " . $file_uri);
+$fp = fopen($file_uri, 'r') or die('could not open file for reading: ' . $file_uri);
+$json_liste = json_decode(fread($fp, filesize($file_uri)), true);
+$json_liste_old_count = count($json_liste);
+fclose($fp);
 
 $file_uri_imdb = plugin_dir_path(__FILE__) . '/imdb_name_basics_small.json';
 $imdb_liste = array();
 
-if (file_exists($file_uri_imdb)) {
-  $fp = fopen($file_uri_imdb, 'r') or die('could not open file for reading: ' . $file_uri_imdb);
-  $imdb_liste = json_decode(fread($fp, filesize($file_uri_imdb)), true);
-  fclose($fp);
-}
+print_debug("reading file " . $file_uri_imdb);
+$fp = fopen($file_uri_imdb, 'r') or die('could not open file for reading: ' . $file_uri_imdb);
+$imdb_liste = json_decode(fread($fp, filesize($file_uri_imdb)), true);
+fclose($fp);
 
 
 $oyuncu_listesi_unique = array_unique($oyuncu_listesi);
 
-if ($debug) {
-  echo "Found " . count($oyuncu_listesi_unique) . " unique oyuncu record\n\n";
-}
+print_debug("Found " . count($oyuncu_listesi_unique) . " oyuncu record we need to work on in DB");
+print_debug(var_export($oyuncu_listesi_unique, true));
 
 $original_has_url = count($json_liste);
 $has_changes = false;
+$oyuncu_added = array();
+$oyuncu_could_not_add = array();
 
 foreach ($oyuncu_listesi_unique as $oyuncu) {
   $oyuncu_index = yirmiiki_shortcode_json_key($oyuncu);
 
   if (!isset($json_liste[$oyuncu_index])) {
-    if ($debug) {
-      echo $oyuncu . "\n";
-    }
+    print_debug("Missing in our data " . $oyuncu);
 
     if (isset($imdb_liste[$oyuncu])) {
       $has_changes = true;
       $found_nmlink = $imdb_liste[$oyuncu];
       $json_liste[$oyuncu_index] = array('link' => "https://www.imdb.com/name/" . $found_nmlink, 'name' => $oyuncu);
+      array_push($oyuncu_added, $oyuncu);
+    } else {
+      array_push($oyuncu_could_not_add, $oyuncu);
     }
   }
 }
@@ -124,16 +138,23 @@ if ($dry_run) {
   fclose($fp);
 }
 
-if (!$force || !$dry_run) {
+if (!$force && !$dry_run) {
   update_option("oyuncu_listesi_son_tarih", $now);
 }
 
 if ($debug) {
-  $total = count($oyuncu_listesi_unique);
-  $has_url = count($json_liste);
-  $remaining = $total - $has_url;
+  $oyuncu_added_count = count($oyuncu_added);
+  $oyuncu_could_not_add_count = count($oyuncu_could_not_add);
+  $json_liste_new_count = count($json_liste);
 
-  echo "\n Total: " . $total . " Old Has URL: " . $original_has_url . " New Has URL: " . $has_url . " Remaining: " . $remaining . "\n\n";
+  print_debug("Added");
+  print_debug(var_export($oyuncu_added, true));
+
+  print_debug("could not add");
+  print_debug(var_export($oyuncu_could_not_add, true));
+
+  print_debug("Stored old count " . $json_liste_old_count . " new count " . $json_liste_new_count . " added " . $oyuncu_added_count . " not added " . $oyuncu_could_not_add_count);
+  echo "\n";
 }
